+++ business-factory/src/interfaces.js
@@ -1,0 +1,75 @@
+export class ExecutiveProductionInterface {
+  async monitor(_workflowId) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ResearchProductionInterface {
+  async createResearchJob(_requestId, _objective, _context) {
+    throw new Error('Not implemented');
+  }
+
+  async executeResearch(_jobId, _request) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class MemoryProductionInterface {
+  recordWorkflowHistory(_payload) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class MetricsProductionInterface {
+  recordWorkflowCompletion(_payload) {
+    throw new Error('Not implemented');
+  }
+
+  retrieveMetrics(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class PerformanceProductionInterface {
+  recordExecution(_payload) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ApprovalProductionInterface {
+  requestApproval(_payload) {
+    throw new Error('Not implemented');
+  }
+
+  isAuthorized(_approvalId) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class CapabilityRegistryProductionInterface {
+  searchCapabilities(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class WorkerOrchestrationProductionInterface {
+  async coordinateWorkflow(_plan) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class ControlCenterProductionInterface {
+  getSystemOverview(_query) {
+    throw new Error('Not implemented');
+  }
+}
+
+export class AtlasInstituteProductionInterface {
+  recordExperiment(_payload) {
+    throw new Error('Not implemented');
+  }
+
+  recommendImprovements(_payload) {
+    throw new Error('Not implemented');
+  }
+}